finally irritated enough to fix some loose ends

+++ patches/utils/xPref.jsm
@@ -1,23 +1,20 @@
-let EXPORTED_SYMBOLS = ['xPref'];
+let EXPORTED_SYMBOLS = ["xPref"];
 
-const {Services} = ChromeUtils.import('resource://gre/modules/Services.jsm');
+const { Services } = ChromeUtils.import("resource://gre/modules/Services.jsm");
 
 var xPref = {
   // Retorna o valor da preferência, seja qual for o tipo, mas não
   // testei com tipos complexos como nsIFile, não sei como detectar
   // uma preferência assim, na verdade nunca vi uma
   get: function (prefPath, def = false, valueIfUndefined, setDefault = true) {
-    let sPrefs = def ?
-                   Services.prefs.getDefaultBranch(null) :
-                   Services.prefs;
+    let sPrefs = def ? Services.prefs.getDefaultBranch(null) : Services.prefs;
 
     try {
       switch (sPrefs.getPrefType(prefPath)) {
         case 0:
           if (valueIfUndefined != undefined)
             return this.set(prefPath, valueIfUndefined, setDefault);
-          else
-            return undefined;
+          else return undefined;
         case 32:
           return sPrefs.getStringPref(prefPath);
         case 64:
@@ -32,16 +29,14 @@ var xPref = {
   },
 
   set: function (prefPath, value, def = false) {
-    let sPrefs = def ?
-                   Services.prefs.getDefaultBranch(null) :
-                   Services.prefs;
+    let sPrefs = def ? Services.prefs.getDefaultBranch(null) : Services.prefs;
 
     switch (typeof value) {
-      case 'string':
+      case "string":
         return sPrefs.setCharPref(prefPath, value) || value;
-      case 'number':
+      case "number":
         return sPrefs.setIntPref(prefPath, value) || value;
-      case 'boolean':
+      case "boolean":
         return sPrefs.setBoolPref(prefPath, value) || value;
     }
     return;
@@ -50,8 +45,7 @@ var xPref = {
   lock: function (prefPath, value) {
     let sPrefs = Services.prefs;
     this.lockedBackupDef[prefPath] = this.get(prefPath, true);
-    if (sPrefs.prefIsLocked(prefPath))
-      sPrefs.unlockPref(prefPath);
+    if (sPrefs.prefIsLocked(prefPath)) sPrefs.unlockPref(prefPath);
 
     this.set(prefPath, value, true);
     sPrefs.lockPref(prefPath);
@@ -62,10 +56,8 @@ var xPref = {
   unlock: function (prefPath) {
     Services.prefs.unlockPref(prefPath);
     let bkp = this.lockedBackupDef[prefPath];
-    if (bkp == undefined)
-      Services.prefs.deleteBranch(prefPath);
-    else
-      this.set(prefPath, bkp, true);
+    if (bkp == undefined) Services.prefs.deleteBranch(prefPath);
+    else this.set(prefPath, bkp, true);
   },
 
   clear: Services.prefs.clearUserPref,
@@ -77,12 +69,12 @@ var xPref = {
   addListener: function (prefPath, trat) {
     this.observer = function (aSubject, aTopic, prefPath) {
       return trat(xPref.get(prefPath), prefPath);
-    }
+    };
 
     Services.prefs.addObserver(prefPath, this.observer);
     return {
       prefPath: prefPath,
-      observer: this.observer
+      observer: this.observer,
     };
   },
 
@@ -90,5 +82,5 @@ var xPref = {
   // Só precisa passar a var definida quando adicionou
   removeListener: function (obs) {
     Services.prefs.removeObserver(obs.prefPath, obs.observer);
-  }
-}
+  },
+};
